+++ ui/src/app/modules/status/widgets/cpu-widget/cpu-widget.component.ts
@@ -1,9 +1,10 @@
-import { Component, OnInit, OnDestroy } from '@angular/core';
+import { Component, OnInit, OnDestroy, ViewChild, ElementRef } from '@angular/core';
 import { Subscription, interval } from 'rxjs';
+import { ChartOptions } from 'chart.js';
+import { Color, BaseChartDirective } from 'ng2-charts';
 
 import { WsService } from '../../../../core/ws.service';
 import { AuthService } from '../../../../core/auth/auth.service';
-
 
 @Component({
   selector: 'app-cpu-widget',
@@ -14,9 +15,53 @@ export class CpuWidgetComponent implements OnInit, OnDestroy {
   private io = this.$ws.getExistingNamespace('status');
   private intervalSubscription: Subscription;
 
+  @ViewChild(BaseChartDirective, { static: true }) private chart: BaseChartDirective;
+  @ViewChild('widgetbackground', { static: true }) private widgetBackground: ElementRef;
+
   public cpu = {} as any;
   public cpuTemperature = {} as any;
-  public currentLoad = {} as any;
+  public currentLoad = 0;
+
+  public lineChartData = [{ data: [] }];
+  public lineChartLabels = [];
+
+  public lineChartOptions: (ChartOptions & { annotation: any }) = {
+    responsive: true,
+    legend: {
+      display: false,
+    },
+    tooltips: {
+      enabled: false,
+    },
+    scales: {
+      xAxes: [
+        {
+          display: false,
+        },
+      ],
+      yAxes: [
+        {
+          display: false,
+          ticks: {
+            max: 100,
+            min: 0,
+          },
+        },
+      ],
+    },
+    annotation: {
+      annotations: [],
+    },
+  };
+
+  public lineChartColors: Color[] = [
+    { // grey
+      backgroundColor: 'rgba(148,159,177,0.2)',
+      borderColor: 'rgba(148,159,177,0.2)',
+      pointRadius: 0,
+      borderWidth: 1,
+    },
+  ];
 
   constructor(
     private $ws: WsService,
@@ -25,25 +70,45 @@ export class CpuWidgetComponent implements OnInit, OnDestroy {
 
   ngOnInit() {
     this.io.connected.subscribe(async () => {
-      this.getServerUptimeInfo();
+      this.getServerCpuInfo();
     });
 
     if (this.io.socket.connected) {
-      this.getServerUptimeInfo();
+      this.getServerCpuInfo();
     }
 
     this.intervalSubscription = interval(9000).subscribe(() => {
       if (this.io.socket.connected) {
-        this.getServerUptimeInfo();
+        this.getServerCpuInfo();
       }
     });
+
+    // lookup the chart color based on the current theme
+    const chartColor = getComputedStyle(this.widgetBackground.nativeElement).backgroundColor;
+    if (chartColor) {
+      this.lineChartColors[0].backgroundColor = chartColor;
+      this.lineChartColors[0].borderColor = chartColor;
+    }
   }
 
-  getServerUptimeInfo() {
+  getServerCpuInfo() {
     this.io.request('get-server-cpu-info').subscribe((data) => {
-      this.cpu = data.cpu;
       this.cpuTemperature = data.cpuTemperature;
       this.currentLoad = data.currentLoad;
+
+      if (!this.lineChartData[0].data.length) {
+        this.lineChartData[0].data = data.cpuLoadHistory;
+        this.lineChartLabels = data.cpuLoadHistory.map(x => 'point');
+      } else {
+        this.lineChartData[0].data.push(data.currentLoad);
+        this.lineChartLabels.push('point');
+
+        if (this.lineChartData[0].data.length > 60) {
+          this.lineChartData[0].data.shift();
+          this.lineChartLabels.shift();
+          this.chart.update();
+        }
+      }
     });
   }